Finalise the code for Housing chatbot
