+++ artificial-intelligence/T2_2025/Mental Health and Smart City Chatbot/Housing chatbot/configs/json_schema.py
@@ -1,5 +1,4 @@
-import json
-from typing import Optional, List, Literal
+from typing import Optional, Literal
 from pydantic import BaseModel, Field, conset
 
 # Schema definition for JSON output
