+++ artificial-intelligence/T2_2025/Mental Health and Smart City Chatbot/Housing chatbot/configs/prompts.py
