+++ artificial-intelligence/T2_2025/Mental Health and Smart City Chatbot/Housing chatbot/get_data.py
@@ -1,0 +1,4 @@
+from utils.utils import API_GTFS
+
+if __name__ == "__main__":
+    API_GTFS.download_and_extract_data()
